Changed play.php to play.html
Thus, there are no more traces of PHP. I've decided not to have menu
options for this game. As an additional plus, I no longer have to deal
with WebMatrix, which sometimes fails to get the most updated version of
my files and is slow whenever my internet connection isn't good.