Gave up on function drawQuotes()
I worked on it for about two or three hours, and I still couldn't figure
it out. Therefore, I've saved my work and wrote some documentation about
it in the file, but the function is unusable. It's quite unfortunate, but
I think anymore time put into that function would make the benefits of
implementing that function not worth the effort. I'll have to deal with
drawQuoteText().