Can draw a chain of quotes poorly
I call this progress. I made the algorithm; I now must make it better,
since in its current state, its unacceptable. The first quote doesn't
appear until the user hits Enter (on a blank quote box, which is not
good). Furthermore, the sound effect doesn't play upon each finishing
reading a quote (by pressing Enter). Lastly, it took my pressing Enter
twice to get from the second to the third quote. However, as I checked
with alerts, the end of the chain of quotes is able to be reached.