Action of Enter key is erased before it's done
That is, before the action of keyboard.enterKeyAction is done (due to the
user's having pressed Enter), the action is deleted. Consequently, those
who set keyboard.enterKeyAction.action needn't worry about calling
keyboard.enterKeyAction.erase() in each action. Of course, I did this
intelligently; the action is deleted in a way so that it is still called,
but it was done in a way so that the user can change the action within the
action. If that's confusing, a peek at the changes in the implementation
of setUpEventHandler() may help.