Made quote bubble shorter
I'm considering representing the quote bubble as an object, so that I can
keep the positions of it and its text in one place. I can make the bubble
taller if necessary, but I don't think I'll ever need to make the host
utter a sentence that takes up three lines.