Put quotesToDraw in the global namespace
To make the syntax for adding a quote better, I also created a function
called 'add'. I understand that this isn't the branch that this change
should've been in, but I happened to think of making this change as I was
about to make an object for the money display.