Settled on a way to edit CanvasStack in middle of quote chain
I don't like this solution, but I can't think of a better one. Thus, I
noted the task of finding a better solution in the todo list as a
considerable task perhaps for later. Anyway, in my solution, editing the
CanvasStack in the middle of the chain requires nesting calls of add() and
deployQuoteChain(). A look at the changes in the code will reveal what I
mean.