Fixed a bug in convertCanvasTextIntoSmallerPieces()
The bug was in that the function allowed a text piece to be too big if the
addition of the most recently added word was the cause. Now, as the
function should have, the function checks whether adding the new word
would make a text piece too big before it adds the new word to the text
piece.