Sound effect doesn't play if user fails trying to move question selector
In doing so, I changed the structure of the four functions that deal with
moving the selector. Because of this change, they thankfully will be
easier to use unit tests with. Before I came up with this idea, I had made
a few untested changes to the function that moves the selector to the
right, and those changes are in this commit.