Making canvases fade in is handled by CanvasStack now
The caller can pass a certain parameter to either CanvasStack.add() or
CanvasStack.set() to make the added canvases fade in with a certain CSS
class. This makes it easier to apply the fade in effect at certain times
rather than on certain canvases. For example, the question-display canvas
should fade in when the million dollar question is presented, but not when
any of the first ten questions are presented; the changes I've made have
made this possible (and I'll deal with that next).