Added list of helpful books to the readme
This is the start of the effort to transfer my work in experimenting with
JavaScript GUI custom types from my other repository to this repository.
Specifically, I had been citing a book that had helped in that
experimentation, and I thus decided that I might as well cite some of the
other books.