In GUI custom types, made it clearer which methods are "overridden"
In doing so, I had to rename one of the methods of GUI.Container, because
that method (formerly named select()) has the same name as an abstract
method in GUI.Component, even though the two methods have different
intents. This had occurred because the C++ versions that I "translated"
did this, and C++ allows it because C++ has method overloading (whereas
JavaScript doesn't).