Added another "loss" sound effect
Now, there are three different "loss" sound effects. One plays if a wrong
answer is processed while the user can still be saved. The second plays if
a wrong answer is processed while the user can't be saved. The third plays
if the wrong answer is given to the million dollar question. In this
commit, I've added the second sound effect's clip.