Finished the title screen
I've properly sized and positioned the images. The images are sized with
the Canvas API. I also changed some of the images so that their
backgrounds are not too distracting. Also, in making these changes, I
improved the code slightly so that the images' urls are not put in two
locations each.